+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 [//]: # (current developments)
 
+* Use `pyzstd` or `compression.zstd` (part of Python 3.14), instead of
+  `zstandard / python-zstandard`, for decompression. create/transmute is tied to
+  the `zstandard` API for now.
+
 ## 0.11.0 (2024-10)
 
 * Add Python 3.12 to test matrix.